core: move tests + import
