+++ gsl_programmation/tests/services/test_simulation_projet_services.py
@@ -1,9 +1,9 @@
 import pytest
 
-from gsl_programmation.models import SimulationProjet
-from gsl_programmation.tests.factories import SimulationProjetFactory
 from gsl_projet.tests.factories import ProjetFactory
+from gsl_simulation.models import SimulationProjet
 from gsl_simulation.services.simulation_projet_service import SimulationProjetService
+from gsl_simulation.tests.factories import SimulationProjetFactory
 
 
 @pytest.fixture
