+++ gsl_simulation/tests/test_tasks.py
@@ -13,11 +13,11 @@ from gsl_demarches_simplifiees.tests.factories import DossierFactory
 from gsl_programmation.tests.factories import (
     DetrEnveloppeFactory,
     DsilEnveloppeFactory,
-    SimulationFactory,
 )
 from gsl_projet.tests.factories import DemandeurFactory, ProjetFactory
 from gsl_simulation.models import SimulationProjet
 from gsl_simulation.tasks import add_enveloppe_projets_to_simulation
+from gsl_simulation.tests.factories import SimulationFactory
 
 
 @pytest.fixture
